Try to handle "codepoints" instead of "characters".
In many places, "char" is not a valid type to handle characters, because it
only supports 16 bits. In order to handle the full range of Unicode characters,
we need to use "int".

This allows characters such as 🙌 or 𐐷 to be treated as single codepoints even
though they are encoded as two "char"s in a C# string.