Try to improve glyph substitution.
Instead of having each substitution table operate on the whole string,
we iterate over characters and call all substitution rules on the
current character. If a replacement was performed, we restart at the
current index.

This allows for the following glyph sequence in Segoe UI Emoji to be
properly substituted: { #1254, #390, #423, #1085 } → { #9151, #1086 }
(using rule 2) → { #9161 } (using rule 1). This is for Unicode sequence
U+1F482 U+200D U+2640 U+1F3FB (💂‍♀🏻).